Split disk-mode indexing out of indexer.c into disk_indexer.c
Move the disk half of the per-document indexing pipeline into its own
translation unit: stageText, bulkStageFields, commitDocument, applyDocTable,
applyTextIndex, bulkApplyFields, applyVectorInserts, and the two DocIdMeta
accessors that only the disk path used.

indexDocumentDisk becomes DiskIndexer_IndexDocument, and the disk branch of
doAssignIds is extracted as DiskIndexer_StageDocument, so indexer.c retains
only the mode dispatch rather than inline disk logic.

Three helpers used by both halves lose `static` and take an Indexer_ prefix
(Indexer_RemoveReplacedDocVectorAndGeometry, Indexer_RemoveOldDocStats,
Indexer_AddNewDocStats), declared alongside entryWantsSuffixTrie in the new
indexer_internal.h. disk_indexer.h carries the two entry points in the other
direction.

Pure move: no behavior change.
